requirements.txt
Flask==2.3.2
pandas==2.2.2
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
openpyxl==3.1.2
